src/test/resources/integration-test/runner-cli/requirements.txt
mysql-connector-python~=8.0.29
PyYAML~=6.0
click~=8.1.3
requests~=2.28.0
tenacity~=8.0.1